+++ dominest/src/components/categories/Complain.tsx
@@ -2,11 +2,11 @@ import React, { useEffect, useState } from "react";
 import "../../app/globals.css";
 import { ComponenComplaints } from "@/style/ComponentStyle";
 import RoomSelector from "@/utils/room/roomnumber";
-import { complainborderList } from "@/utils/border/borderlist";
-import { Button, Containerright } from "@/style/InputStyle";
 import "../../app/globals.css";
-import { Container, Table, ButtonContainer } from "@/style/border";
 import Link from "next/link";
+import ComplainList from "@/components/complain/ComplainList";
+import ComplainSelect from "../complain/ComplainSelect";
+import { Container } from "@/style/border";
 interface Post {
   id: number;
   title: string;
@@ -18,35 +18,18 @@ interface Post {
   complaintCause: string;
   complaintResolution: string;
   processState: string;
-  createdBy: string;
+  name: string;
   date: string;
 }
 
 export default function Complaints(props: any) {
   const [posts, setPosts] = useState<Post[]>([]);
   const [currentPage, setCurrentPage] = useState(1);
-  const [totalPages, setTotalPages] = useState(1);
-  const [categoryName, setCategoryName] = useState("");
+  const [selectedRoom, setSelectedRoom] = useState(0);
   const [status, setStatus] = useState("민원 조회");
   const urlLink = `/complain/complainupload/${props.idname[0]}`;
   const [serch, setSerch] = useState("");
   const [statusresult, setStatusresult] = useState("처리중");
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const response = await complainborderList(props.idname, currentPage);
-        console.log(response.data.data);
-        setPosts(response.data.data.complaints);
-        setCurrentPage(response.data.data.page.currentPage);
-        setTotalPages(response.data.data.page.totalPages);
-        setCategoryName(response.data.data.category.categoryName);
-      } catch (error) {
-        console.error("에러 발생", error);
-      }
-    };
-
-    fetchData();
-  }, [currentPage]);
 
   const onChangeStatus = (e: React.ChangeEvent<HTMLSelectElement>) => {
     setStatus(e.target.value);
@@ -57,7 +40,9 @@ export default function Complaints(props: any) {
   const onChangeStatusresult = (e: React.ChangeEvent<HTMLSelectElement>) => {
     setStatusresult(e.target.value);
   };
-
+  const handleRoomChange = (roomNumber: number) => {
+    setSelectedRoom(roomNumber);
+  };
   return (
     <div>
       <ComponenComplaints>
@@ -67,18 +52,21 @@ export default function Complaints(props: any) {
         <div>
           <div>
             <select onChange={onChangeStatus}>
+              <option value="전체 조회">전체 조회</option>
               <option value="민원 조회">민원 조회</option>
               <option value="호실 조회">호실 조회</option>
               <option value="처리 조회">처리 조회</option>
             </select>
-            {status === "민원 조회" ? (
+            {status === "전체 조회" ? (
+              <div></div>
+            ) : status === "민원 조회" ? (
               <input
                 type="text"
                 placeholder="검색어를 입력하세요"
                 onChange={onChangeSerch}
               />
             ) : status === "호실 조회" ? (
-              <RoomSelector />
+              <RoomSelector onRoomChange={handleRoomChange} />
             ) : (
               <select onChange={onChangeStatusresult}>
                 <option value="처리중">처리중</option>
@@ -93,63 +81,17 @@ export default function Complaints(props: any) {
           </div>
         </div>
       </ComponenComplaints>
-      <Container>
-        <h3>{categoryName}</h3>
-        <Table>
-          <thead>
-            <tr>
-              <th>번호</th>
-              <th>방번호</th>
-              <th>민원 내역</th>
-              <th>민원 답변</th>
-              <th>민원 결과</th>
-              <th>민원인</th>
-              <th>작성일</th>
-            </tr>
-          </thead>
-          <tbody>
-            {posts.length > 0 ? (
-              posts.map((post) => (
-                <tr key={post.id}>
-                  <td>{post.id}</td>
-                  <td>{post.roomNo}</td>
-                  <td>{post.complaintCause}</td>
-                  <td>{post.complaintResolution}</td>
-                  <td
-                    style={{
-                      color: post.processState === "처리중" ? "red" : "black",
-                      fontWeight:
-                        post.processState === "처리중" ? "bold" : "normal",
-                    }}
-                  >
-                    {post.processState}
-                  </td>
-                  <td>{post.name}</td>
-                  <td>{post.date}</td>
-                </tr>
-              ))
-            ) : (
-              <tr>
-                <td colSpan={4}>데이터가 없습니다.</td>
-              </tr>
-            )}
-          </tbody>
-        </Table>
 
-        {/* 페이지 넘기기 */}
-        <ButtonContainer>
-          {currentPage > 1 && (
-            <Button onClick={() => setCurrentPage(currentPage - 1)}>
-              이전 페이지
-            </Button>
-          )}
-          {currentPage < totalPages && (
-            <Button onClick={() => setCurrentPage(currentPage + 1)}>
-              다음 페이지
-            </Button>
-          )}
-        </ButtonContainer>
-      </Container>
+      {status === "전체 조회" && <ComplainList idname={props.idname[0]} />}
+      {status === "민원 조회" && <ComplainList idname={props.idname[0]} />}
+      {selectedRoom !== 0 && status === "호실 조회" && (
+        <Container>
+          <ComplainSelect
+            idname={props.idname[0]}
+            selectedRoom={selectedRoom}
+          />
+        </Container>
+      )}
     </div>
   );
 }
