+++ dominest/src/pages/complain/complainupload/[id].tsx
@@ -49,6 +49,7 @@ export default function ComplainUpload() {
   const handleDateChange = (event: { target: { value: any } }) => {
     setDate(event.target.value);
   };
+
   const complainUpload = async () => {
     if (selectedRoom === 0) {
       return alert("방번호를 선택해주세요");
